Add daily limit for deletions

+++ bot/AVHandler.go
@@ -5,21 +5,69 @@ import (
 	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api"
 	"log"
 	"math/rand"
+	"time"
 )
 
 /*
 	AV - Audio/Video
 */
 
-var responses = [3]string{
-	"Пиши буковами, %s",
-	"Я запрещаю вам срать, %s",
-	"Президент заборонив срати, %s",
+const defaultLimit = 3
+var (
+	responses = [3]string{
+		"Пиши буковами, %s",
+		"Я запрещаю вам срать, %s",
+		"Президент заборонив срати, %s",
+	}
+	cache = make(map[int64]map[int]Cache)
+)
+type Cache struct {
+	time time.Time
+	sendVoice int64
 }
+
+
+func updateCache(chatID int64, update tgbotapi.Update) {
+	if _, ok := cache[chatID]; !ok {
+		cache[chatID] = map[int]Cache{
+			update.Message.From.ID: {
+				time.Now(),
+				1,
+			},
+		}
+	} else {
+		if _, ok := cache[chatID][update.Message.From.ID]; !ok {
+			cache[chatID][update.Message.From.ID] = Cache{
+				time.Now(),
+				1,
+			}
+		} else {
+			cached := cache[chatID][update.Message.From.ID]
+			if time.Now().Sub(cached.time).Hours() >= 24 {
+				cache[chatID][update.Message.From.ID] = Cache{
+					time.Now(),
+					1,
+				}
+			} else if cached.sendVoice <= defaultLimit {
+				cache[chatID][update.Message.From.ID] = Cache{
+					time.Now(),
+					cached.sendVoice + 1,
+				}
+			}
+		}
+	}
+}
+
 
 func HandleAV(bot *tgbotapi.BotAPI, update tgbotapi.Update) {
 	chatID := update.Message.Chat.ID
 	messageID := update.Message.MessageID
+
+	updateCache(chatID, update)
+
+	if cache[chatID][update.Message.From.ID].sendVoice <= defaultLimit {
+		return
+	}
 
 	_, err := bot.DeleteMessage(tgbotapi.DeleteMessageConfig{
 		ChatID: chatID,
